[DSCP-454] Split `Dscp.DB.SQL.Util` into submodules
